edited bottom nav labels
